requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
ta-lib>=0.4.0
loguru>=0.6.0
pyyaml>=6.0.0